+++ regress/liboverride-test.c
@@ -66,14 +66,15 @@ main(int argc, const char *argv[]) {
     
     if (getenv("LIBOVERRIDE_SET") == NULL) {
         char *cwd = getcwd(NULL, 0);
-        char *so = (char *)malloc(strlen(cwd) + 64);
+        size_t so_size = strlen(cwd) + 64;
+        char *so = (char *)malloc(so_size);
         if (so == NULL) {
             if (verbose) {
                 printf("malloc failed\n");
             }
             exit(2);
         }
-        sprintf(so, "%s/libliboverride.so", cwd);
+        snprintf(so, so_size, "%s/libliboverride.so", cwd);
         setenv("LIBOVERRIDE_SET", "1", 1);
         setenv("LD_PRELOAD", so, 1);
         execv(argv[0], (void *)argv);