+++ src/zipmerge.c
@@ -80,6 +80,8 @@ int confirm;
 zip_flags_t name_flags;
 
 static int confirm_replace(zip_t *, const char *, zip_uint64_t, zip_t *, const char *, zip_uint64_t);
+static void copy_extra_fields(zip_t *destination_archive, zip_uint64_t destination_index, zip_t *source_archive, zip_uint64_t source_index, zip_flags_t flags);
+static int copy_file(zip_t *destination_archive, zip_int64_t destination_index, zip_t *source_archive, zip_uint64_t source_index, const char* name);
 static zip_t *merge_zip(zip_t *, const char *, const char *);
 
 
@@ -216,7 +218,6 @@ confirm_replace(zip_t *za, const char *tname, zip_uint64_t it, zip_t *zs, const 
 static zip_t *
 merge_zip(zip_t *za, const char *tname, const char *sname) {
     zip_t *zs;
-    zip_source_t *source;
     zip_int64_t ret, idx;
     zip_uint64_t i;
     int err;
@@ -244,8 +245,7 @@ merge_zip(zip_t *za, const char *tname, const char *sname) {
                 break;
 
             case 1:
-                if ((source = zip_source_zip(za, zs, i, 0, 0, 0)) == NULL || zip_replace(za, (zip_uint64_t)idx, source) < 0) {
-                    zip_source_free(source);
+                if (copy_file(za, idx, zs, i, NULL) < 0) {
                     fprintf(stderr, "%s: cannot replace '%s' in `%s': %s\n", progname, fname, tname, zip_strerror(za));
                     zip_close(zs);
                     return NULL;
@@ -266,8 +266,7 @@ merge_zip(zip_t *za, const char *tname, const char *sname) {
             }
         }
         else {
-            if ((source = zip_source_zip(za, zs, i, 0, 0, 0)) == NULL || zip_add(za, fname, source) < 0) {
-                zip_source_free(source);
+            if (copy_file(za, -1, zs, i, fname) < 0) {
                 fprintf(stderr, "%s: cannot add '%s' to `%s': %s\n", progname, fname, tname, zip_strerror(za));
                 zip_close(zs);
                 return NULL;
@@ -276,4 +275,50 @@ merge_zip(zip_t *za, const char *tname, const char *sname) {
     }
 
     return zs;
+}
+
+
+static int copy_file(zip_t *destination_archive, zip_int64_t destination_index, zip_t *source_archive, zip_uint64_t source_index, const char* name) {
+    zip_source_t *source = zip_source_zip(destination_archive, source_archive, source_index, 0, 0, 0);
+
+    if (source == NULL) {
+        return -1;
+    }
+
+    if (destination_index >= 0) {
+        if (zip_replace(destination_archive, (zip_uint64_t)destination_index, source) < 0) {
+            zip_source_free(source);
+            return -1;
+        }
+    }
+    else {
+        destination_index = zip_file_add(destination_archive, name, source, 0);
+        if (destination_index < 0) {
+            zip_source_free(source);
+            return -1;
+        }
+    }
+
+    copy_extra_fields(destination_archive, (zip_uint64_t)destination_index, source_archive, source_index, ZIP_FL_CENTRAL);
+    copy_extra_fields(destination_archive, (zip_uint64_t)destination_index, source_archive, source_index, ZIP_FL_LOCAL);
+
+    return 0;
+}
+
+
+static void copy_extra_fields(zip_t *destination_archive, zip_uint64_t destination_index, zip_t *source_archive, zip_uint64_t source_index, zip_flags_t flags) {
+    zip_int16_t n;
+    zip_uint16_t i, id, length;
+    const zip_uint8_t *data;
+
+    if ((n = zip_file_extra_fields_count(source_archive, source_index, flags)) < 0) {
+        return;
+    }
+
+    for (i = 0; i < n; i++) {
+        if ((data = zip_file_extra_field_get(source_archive, source_index, i, &id, &length, flags)) == NULL) {
+            continue;
+        }
+        zip_file_extra_field_set(destination_archive, destination_index, id, ZIP_EXTRA_FIELD_NEW, data, length, flags);
+    }
 }